Add interface and methods
