+++ pkg/odo/cli/component/create_refactor.go
@@ -5,18 +5,10 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strings"
 
-	"k8s.io/klog"
-
-	registryLibrary "github.com/devfile/registry-support/registry-library/library"
-	registryConsts "github.com/openshift/odo/pkg/odo/cli/registry/consts"
-	"github.com/openshift/odo/pkg/preference"
+	scontext "github.com/openshift/odo/pkg/segment/context"
 
 	"github.com/openshift/odo/pkg/catalog"
-	"github.com/openshift/odo/pkg/component"
-	"github.com/openshift/odo/pkg/odo/cli/component/ui"
-
 	registryUtil "github.com/openshift/odo/pkg/odo/cli/registry/util"
 	"github.com/zalando/go-keyring"
 
@@ -35,127 +27,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
-
-// Get Metadata
-// If the devfile does not exist, then check and Validate Registry
-// Fetch Devfile
-// If devfile was manually created, rollback
-
-type CreateMethod interface {
-	SetMetadata(co *CreateOptions, cmd *cobra.Command, catalogDevfileList catalog.DevfileComponentTypeList, args []string) error
-	//CheckAndValidateRegistry(registryName string) (catalog.DevfileComponentTypeList, error)
-	FetchDevfile(co *CreateOptions, catalogDevfileList catalog.DevfileComponentTypeList) error
-	//Rollback(devfilePath string)
-}
-
-type InteractiveCreateMethod struct{}
-
-func (icm *InteractiveCreateMethod) SetMetadata(co *CreateOptions, cmd *cobra.Command, catalogDevfileList catalog.DevfileComponentTypeList, args []string) error {
-	var err error
-	// Component type: We provide devfile component list to let user choose
-	componentType := ui.SelectDevfileComponentType(catalogDevfileList.Items)
-
-	// Component name: User needs to specify the component name, by default it is component type that user chooses
-	componentName := ui.EnterDevfileComponentName(componentType)
-
-	// Component namespace: User needs to specify component namespace, by default it is the current active namespace
-	var componentNamespace string
-	if cmd.Flags().Changed("project") {
-		componentNamespace, err = cmd.Flags().GetString("project")
-		if err != nil {
-			return err
-		}
-	} else {
-		client, e := genericclioptions.Client()
-		// if the user is logged in or if we have cluster information, display the default project
-		if e == nil {
-			componentNamespace = ui.EnterDevfileComponentProject(client.GetCurrentProjectName())
-		}
-	}
-
-	co.devfileMetadata.componentType = componentType
-	co.devfileName = componentType
-	co.devfileMetadata.componentName = componentName
-	co.devfileMetadata.componentNamespace = componentNamespace
-
-	return err
-}
-
-func (icm *InteractiveCreateMethod) FetchDevfile(co *CreateOptions, catalogDevfileList catalog.DevfileComponentTypeList) error {
-	return fetchDevfile(co, catalogDevfileList)
-}
-
-type DirectCreateMethod struct{}
-
-func (dcm *DirectCreateMethod) SetMetadata(co *CreateOptions, cmd *cobra.Command, catalogDevfileList catalog.DevfileComponentTypeList, args []string) error {
-	var err error
-	componentType := args[0]
-
-	co.devfileMetadata.componentType = componentType
-	co.devfileName = componentType
-	var componentName string
-	if len(args) == 2 {
-		componentName = args[1]
-	} else {
-		componentName, err = createDefaultComponentName(
-			componentType,
-			co.componentContext,
-		)
-		if err != nil {
-			return err
-		}
-	}
-	co.devfileMetadata.componentName = componentName
-	return err
-
-}
-
-func (dcm *DirectCreateMethod) FetchDevfile(co *CreateOptions, catalogDevfileList catalog.DevfileComponentTypeList) error {
-	return fetchDevfile(co, catalogDevfileList)
-}
-
-type HTTPCreateMethod struct{}
-
-func (hcm HTTPCreateMethod) FetchDevfile(co *CreateOptions, catalogDevfileList catalog.DevfileComponentTypeList) error {
-	devfileSpinner := log.Spinnerf("Creating a devfile component from devfile path: %s", co.devfileMetadata.devfilePath.value)
-	defer devfileSpinner.End(false)
-
-	params := util.HTTPRequestParams{
-		URL:   co.devfileMetadata.devfilePath.value,
-		Token: co.devfileMetadata.token,
-	}
-	devfileData, err := util.DownloadFileInMemory(params)
-	if err != nil {
-		return errors.Wrapf(err, "failed to download devfile for devfile component from %s", co.devfileMetadata.devfilePath.value)
-	}
-	err = ioutil.WriteFile(co.DevfilePath, devfileData, 0644) // #nosec G306
-	if err != nil {
-		return errors.Wrapf(err, "unable to save devfile to %s", co.DevfilePath)
-	}
-	devfileSpinner.End(true)
-	return nil
-}
-
-type FileCreateMethod struct{}
-
-func (fcm FileCreateMethod) FetchDevfile(co *CreateOptions, catalogDevfileList catalog.DevfileComponentTypeList) error {
-	devfileAbsolutePath, err := filepath.Abs(co.devfileMetadata.devfilePath.value)
-	if err != nil {
-		return err
-	}
-	devfileSpinner := log.Spinnerf("Creating a devfile component from devfile path: %s", devfileAbsolutePath)
-	defer devfileSpinner.End(false)
-	devfileData, err := ioutil.ReadFile(co.devfileMetadata.devfilePath.value)
-	if err != nil {
-		return errors.Wrapf(err, "failed to read devfile from %s", co.devfileMetadata.devfilePath)
-	}
-	err = ioutil.WriteFile(co.DevfilePath, devfileData, 0644) // #nosec G306
-	if err != nil {
-		return errors.Wrapf(err, "unable to save devfile to %s", co.DevfilePath)
-	}
-	devfileSpinner.End(true)
-	return nil
-}
 
 func (co *CreateOptions) Complete(name string, cmd *cobra.Command, args []string) (err error) {
 	// GETTERS
@@ -233,111 +104,64 @@ func (co *CreateOptions) Complete(name string, cmd *cobra.Command, args []string
 	}
 
 	log.Info("Devfile Object Creation")
+	var catalogDevfileList catalog.DevfileComponentTypeList
+	var createMethod CreateMethod
 	switch {
 	case co.devfileMetadata.userCreatedDevfile:
+		createMethod = UserCreatedDevfileMethod{}
+		err = createMethod.SetMetadata(co, cmd, args, catalogDevfileList)
+		if err != nil {
+			return err
+		}
 	case co.devfileMetadata.devfilePath.value != "":
-	case co.interactive:
-	default:
-
-	}
-	//--------------------------------------------------------------------------------------------------------------------------------------------
-	//	Existing devfile Mode; co.devfileName = ""
-	if co.devfileMetadata.userCreatedDevfile {
-		devfileAbsolutePath, err := filepath.Abs(co.DevfilePath)
-		if err != nil {
-			return err
-		}
-		devfileSpinner := log.Spinnerf("Creating a devfile component from devfile path %s", devfileAbsolutePath)
-		defer devfileSpinner.End(true)
-
-		// Set the component name
-		co.devfileMetadata.componentName, err = getComponentNameForExistingDevfile(co.DevfilePath, args)
-		if err != nil {
-			return err
-		}
-
-		// Set the component type
-		co.devfileMetadata.componentType, err = getComponentType(co.DevfilePath)
-		if err != nil {
-			return err
-		}
-		return nil
-	}
-
-	//--------------------------------------------------------------------------------------------------------------------------------------------
-	//--devfile Mode; co.devfileName = ""
-	if co.devfileMetadata.devfilePath.value != "" {
+		//--devfile Mode; co.devfileName = ""
 		fileErr := util.ValidateFile(co.devfileMetadata.devfilePath.value)
 		urlErr := util.ValidateURL(co.devfileMetadata.devfilePath.value)
 		if fileErr != nil && urlErr != nil {
 			return errors.Errorf("the devfile path you specify is invalid with either file error %q or url error %q", fileErr, urlErr)
 		} else if fileErr == nil {
 			co.devfileMetadata.devfilePath.protocol = "file"
-			var fcm = FileCreateMethod{}
-			err = fcm.FetchDevfile(co, catalog.DevfileComponentTypeList{})
-			if err != nil {
-				rollback(co.DevfilePath)
-				return err
-			}
+			createMethod = FileCreateMethod{}
+
 		} else if urlErr == nil {
 			co.devfileMetadata.devfilePath.protocol = "http(s)"
-			var hcm = HTTPCreateMethod{}
-			err = hcm.FetchDevfile(co, catalog.DevfileComponentTypeList{})
-			if err != nil {
-				rollback(co.DevfilePath)
-				return err
-			}
+			createMethod = HTTPCreateMethod{}
 		}
 
-		// get the custom component name
-		co.devfileMetadata.componentName, err = getComponentNameForExistingDevfile(co.DevfilePath, args)
+		err = createMethod.FetchDevfileAndCreateComponent(co, catalogDevfileList)
 		if err != nil {
 			rollback(co.DevfilePath)
 			return err
 		}
-		co.devfileMetadata.componentType, err = getComponentType(co.DevfilePath)
+		err = createMethod.SetMetadata(co, cmd, args, catalogDevfileList)
 		if err != nil {
 			rollback(co.DevfilePath)
 			return err
 		}
-		return nil
-	}
+	default:
+		if co.interactive {
+			createMethod = InteractiveCreateMethod{}
+		} else {
+			createMethod = DirectCreateMethod{}
+		}
 
-	//Interactive Mode
-	if co.interactive {
-		var icm = InteractiveCreateMethod{}
-		catalogDevfileList, err := validateAndFetchRegistry(co.devfileMetadata.devfileRegistry.Name)
+		catalogDevfileList, err = validateAndFetchRegistry(co.devfileMetadata.devfileRegistry.Name)
 		if err != nil {
 			return err
 		}
-		err = icm.SetMetadata(co, cmd, catalogDevfileList, args)
+		err = createMethod.SetMetadata(co, cmd, args, catalogDevfileList)
 		if err != nil {
 			return err
 		}
-		err = icm.FetchDevfile(co, catalogDevfileList)
+		err = createMethod.FetchDevfileAndCreateComponent(co, catalogDevfileList)
 		if err != nil {
 			rollback(co.DevfilePath)
 			return err
 		}
-		return nil
 	}
-	//Normal Mode
-	// Component type: Get from full command's first argument (mandatory in direct mode)
-	var dcm = DirectCreateMethod{}
-	err = dcm.SetMetadata(co, cmd, catalog.DevfileComponentTypeList{}, args)
-	if err != nil {
-		return err
-	}
+	// Adding user provided devfile name to telemetry data
+	scontext.SetDevfileName(cmd.Context(), co.devfileName)
 
-	catalogDevfileList, err := validateAndFetchRegistry(co.devfileMetadata.devfileRegistry.Name)
-	if err != nil {
-		return err
-	}
-	err = dcm.FetchDevfile(co, catalogDevfileList)
-	if err != nil {
-		rollback(co.DevfilePath)
-		return err
-	}
 	return nil
 }
 
@@ -367,6 +191,9 @@ func (co *CreateOptions) Validate() (err error) {
 }
 
 func (co *CreateOptions) Run(cmd *cobra.Command) (err error) {
+	// Adding component type to telemetry data
+	scontext.SetComponentType(cmd.Context(), co.devfileMetadata.componentType)
+
 	devObj, err := devfileParseFromFile(co.DevfilePath, false)
 	if err != nil {
 		return errors.New("Failed to parse the devfile")
@@ -441,144 +268,6 @@ func (co *CreateOptions) Run(cmd *cobra.Command) (err error) {
 	}
 
 	return nil
-}
-
-func validateAndFetchRegistry(registryName string) (catalog.DevfileComponentTypeList, error) {
-	// Validate if the component type is available
-	if registryName != "" {
-		registryExistSpinner := log.Spinnerf("Checking if the registry %q exists", registryName)
-		defer registryExistSpinner.End(false)
-		registryList, e := catalog.GetDevfileRegistries(registryName)
-		if e != nil {
-			return catalog.DevfileComponentTypeList{}, errors.Wrap(e, "failed to get registry")
-		}
-		if len(registryList) == 0 {
-			return catalog.DevfileComponentTypeList{}, errors.Errorf("registry %s doesn't exist, please specify a valid registry via --registry", registryName)
-		}
-		registryExistSpinner.End(true)
-	}
-
-	klog.V(4).Infof("Fetching the available devfile components")
-	// Get available devfile components for checking devfile compatibility
-	catalogDevfileList, err := catalog.ListDevfileComponents(registryName)
-	if err != nil {
-		return catalog.DevfileComponentTypeList{}, err
-	}
-
-	if registryName != "" && catalogDevfileList.Items == nil {
-		return catalog.DevfileComponentTypeList{}, errors.Errorf("can't create devfile component from registry %s", registryName)
-	}
-
-	if len(catalogDevfileList.DevfileRegistries) == 0 {
-		return catalog.DevfileComponentTypeList{}, errors.New("Registry is empty, please run `odo registry add <registry name> <registry URL>` to add a registry\n")
-	}
-	return catalogDevfileList, nil
-}
-
-func fetchDevfile(co *CreateOptions, catalogDevfileList catalog.DevfileComponentTypeList) error {
-	hasComponent := false
-	var devfileExistSpinner *log.Status
-	if co.devfileMetadata.devfileRegistry.Name != "" {
-		devfileExistSpinner = log.Spinnerf("Checking if the devfile for %q exists on registry %q", co.devfileMetadata.componentType, co.devfileMetadata.devfileRegistry.Name)
-	} else {
-		devfileExistSpinner = log.Spinnerf("Checking if the devfile for %q exists on available registries", co.devfileMetadata.componentType)
-	}
-	defer devfileExistSpinner.End(false)
-
-	for _, devfileComponent := range catalogDevfileList.Items {
-		if co.devfileMetadata.componentType == devfileComponent.Name {
-			hasComponent = true
-			co.devfileMetadata.devfileLink = devfileComponent.Link
-			co.devfileMetadata.devfileRegistry = devfileComponent.Registry
-			break
-		}
-	}
-	if hasComponent {
-		devfileExistSpinner.End(true)
-	} else {
-		devfileExistSpinner.End(false)
-		return fmt.Errorf("devfile component type %q is not supported, please run `odo catalog list components` for a list of supported devfile component types", co.devfileMetadata.componentType)
-
-	}
-
-	registrySpinner := log.Spinnerf("Creating a devfile component from registry %q", co.devfileMetadata.devfileRegistry.Name)
-	if registryUtil.IsGitBasedRegistry(co.devfileMetadata.devfileRegistry.URL) {
-		registryUtil.PrintGitRegistryDeprecationWarning()
-	}
-
-	// Download devfile from registry
-	var params util.HTTPRequestParams
-	// For GitHub based registries
-	if strings.Contains(co.devfileMetadata.devfileRegistry.URL, "github") {
-		params = util.HTTPRequestParams{
-			URL: co.devfileMetadata.devfileRegistry.URL + co.devfileMetadata.devfileLink,
-		}
-
-		secure, e := registryUtil.IsSecure(co.devfileMetadata.devfileRegistry.Name)
-		if e != nil {
-			return e
-		}
-
-		if secure {
-			var token string
-			token, err := keyring.Get(fmt.Sprintf("%s%s", util.CredentialPrefix, co.devfileMetadata.devfileRegistry.Name), registryUtil.RegistryUser)
-			if err != nil {
-				return errors.Wrap(err, "unable to get secure registry credential from keyring")
-			}
-			params.Token = token
-		}
-
-		cfg, err := preference.New()
-		if err != nil {
-			return err
-		}
-		devfileData, err := util.DownloadFileInMemoryWithCache(params, cfg.GetRegistryCacheTime())
-		if err != nil {
-			return err
-		}
-
-		err = ioutil.WriteFile(co.DevfilePath, devfileData, 0644) // #nosec G306
-		if err != nil {
-			return err
-		}
-	} else {
-		err := registryLibrary.PullStackFromRegistry(co.devfileMetadata.devfileRegistry.URL, co.devfileMetadata.componentType, co.componentContext, false, registryConsts.TelemetryClient)
-		if err != nil {
-			return err
-		}
-		// if the function fails, remove this newly created devfile
-		defer func() {
-			if err != nil {
-				os.Remove(co.DevfilePath)
-			}
-		}()
-	}
-	registrySpinner.End(true)
-	return nil
-}
-
-func getComponentNameForExistingDevfile(devfilePath string, args []string) (string, error) {
-	// get the custom component name
-	if len(args) > 0 {
-		return args[0], nil
-	}
-	devObj, err := devfileParseFromFile(devfilePath, false)
-	if err != nil {
-		currentDirPath, err := os.Getwd()
-		if err != nil {
-			return "", err
-		}
-		return filepath.Base(currentDirPath), nil
-	}
-	return devObj.GetMetadataName(), nil
-}
-
-func getComponentType(devfilePath string) (string, error) {
-	devObj, err := devfileParseFromFile(devfilePath, false)
-	if err != nil {
-		return "", err
-	}
-	return component.GetComponentTypeFromDevfileMetadata(devObj.Data.GetMetadata()), nil
 }
 
 // Rollback in case a devfile was manually created.
